Update component to fetch with App
