+++ src/components/AppleContainer.js
@@ -1,22 +1,14 @@
 import React, { useState, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchAppleData } from '../redux/apple/appleSlice';
+import { useSelector } from 'react-redux';
 import DataHandler from './DataHandler';
 import Header from './Header';
 
 const Apple = () => {
   const { apple, isLoading, error } = useSelector((state) => state.apple);
-  const dispatch = useDispatch();
   const [isSearchVisible, setIsSearchVisible] = useState(false);
   const [searchQuery, setSearchQuery] = useState('');
   const [filteredData, setFilteredData] = useState([]);
   const [resultCount, setResultCount] = useState(0);
-
-  useEffect(() => {
-    if (apple.length === 0) {
-      dispatch(fetchAppleData());
-    }
-  }, [dispatch, apple]);
 
   useEffect(() => {
     if (searchQuery !== '') {
