+++ src/components/Home.js
@@ -1,31 +1,38 @@
+import { useSelector } from 'react-redux';
 import PageLinks from './PageLinks';
 import Header from './Header';
 import '../styles/HomePage.css';
 
-const Home = () => (
-  <>
-    <Header />
-    <main className="home-container">
-      <div className="landing">
-        <h1>
-          Hello! This is Mobile Tech Financial Digest.
-        </h1>
-        <p>
-          We&apos;ll be taking a look at some companies that have a large presence
-          in the smartphone manufacturing market and their numbers.
-          Note that all these companies are on public stock exchanges and are ordered randomly.
-        </p>
-        <p>
-          Below lies a list of companies to choose from.
-          Select one to see their numbers for up to the last 5 years.
-        </p>
-      </div>
-      <div className="companies">
-        <p className="companies-found">(4) Companies available</p>
-        <PageLinks />
-      </div>
-    </main>
-  </>
-);
+const Home = () => {
+  const { apple } = useSelector((state) => state.apple);
+  const { samsung } = useSelector((state) => state.samsung);
+  const { xiaomi } = useSelector((state) => state.xiaomi);
+  const { google } = useSelector((state) => state.google);
+  return (
+    <>
+      <Header />
+      <main className="home-container">
+        <div className="landing">
+          <h1>
+            Hello! This is Mobile Tech Financial Digest.
+          </h1>
+          <p>
+            We&apos;ll be taking a look at some companies that have a large presence
+            in the smartphone manufacturing market and their numbers.
+            Note that all these companies are on public stock exchanges and are ordered randomly.
+          </p>
+          <p>
+            Below lies a list of companies to choose from.
+            Select one to see their numbers for up to the last 5 years.
+          </p>
+        </div>
+        <div className="companies">
+          <p className="companies-found">(4) Companies available</p>
+          <PageLinks apple={apple[0]} xiaomi={xiaomi[0]} google={google[0]} samsung={samsung[0]} />
+        </div>
+      </main>
+    </>
+  );
+};
 
 export default Home;
